+++ main.ts
@@ -33,7 +33,7 @@ import { createFileclassApi, FileclassApi } from "./src/api/fileclassApi";
 import { CanvasEngine } from "./src/fields/canvas/canvasEngine";
 import { FieldIndicator } from "./src/ui/indicator/fieldIndicator";
 import { LinkIndicator } from "./src/ui/indicator/linkIndicator";
-import { applyDraggableModals } from "./src/ui/modalDrag";
+import { applyDraggableModals, applyShorterModals } from "./src/ui/modalDrag";
 import { registerPrimaryActionShortcut } from "./src/ui/primaryAction";
 import { PropertyEditButtons } from "./src/ui/propertyEditButtons";
 import { NoteFieldsModal } from "./src/ui/noteFieldsModal";
@@ -88,6 +88,8 @@ export default class FileclassPlugin extends Plugin {
 		// body class, and it goes away with the plugin.
 		applyDraggableModals(this.settings.enableDraggableModals);
 		this.register(() => applyDraggableModals(false));
+		applyShorterModals(!!this.settings.shorterModal);
+		this.register(() => applyShorterModals(false));
 		this.addChild(new FileclassContextMenu(this));
 		this.indicator = this.addChild(new FieldIndicator(this));
 		this.linkIndicator = this.addChild(new LinkIndicator(this));